feat: More re-organising for Poller
