+++ Sources/UnleashProxyClientSwift/Poller/PollerError.swift
@@ -1,0 +1,7 @@
+public enum PollerError: Error {
+    case decoding
+    case network
+    case url
+    case noResponse
+    case unhandledStatusCode
+}
