+++ Sources/UnleashProxyClientSwift/Poller/PollerSession.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol PollerSession {
+    func perform(_ request: URLRequest, completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void)
+}
