+++ Sources/UnleashProxyClientSwift/Poller/StorageProvider.swift
@@ -1,0 +1,5 @@
+public protocol StorageProvider {
+    func set(value: Toggle?, key: String)
+    func value(key: String) -> Toggle?
+    func clear()
+}
